Now the realm appears in the URL
